Improve styling for sidebar
This isn't perfect but it fixes an issue where the sidebar is
_frequently_ broken.
What was happening was that previously, the sidebar would not have the
height of a full page (this is particularly common when the user has
only a few saved recipes.
Now, the height is at least 100% of the view height.

This does introduce a _different_ issue where 100vh includes the 72px
for the header bar, which means even on very short pages, the site has a
scroll.
That said, in common use, this will be a more rare occurrence and so
it's acceptable at this time.