Swap out getRecipe for getRecipeWith...
Likely will want to think about this pattern more since the point is
that the client will likely *always* want to get a joined recipe.

Are there reasons to build the APIs (now) for the more specific tables -
likely not.
Doing so feels like busy work and not toward the purpose of the actual
app.