Add API routes for auth and pw change
Adding the concept of a role to the user model, we now have users and
admins.

Additionally, there are now some basic API routes:
1. /api/auth
1. /api/protected/user/$userid/password

The former allows a user to authenticate and get a JWT
The latter is to be able to change a user's password.

This involved some database migration as well as setting up JWTs and a
few other utilities.